requirements.txt
pandas>=1.5.0
numpy>=1.23.0
python-dotenv>=1.0.0
pyarrow>=12.0.0
pytest>=7.4.0
xgboost>=2.0.0
scikit-learn>=1.3.0
requests>=2.31.0
MetaTrader5>=5.0.45; sys_platform == "win32"
